Basic sass/scss compilation works as a service
Using post requests to /scss and /sass, this service will compile
and reutrn valid css. I'd like ot get this working with compass, but I
have yet figured out how that will work with Compass' image processing
capabilities. It may be that sassd will have to run on a server with a
shared file-system and, instead of accepting sass/scss strings as input,
it accepts a file path. If I go down that road, we should probably just
use sockets instead of a full HTTP stack. Not sure yet.